refactor: router 리팩토링
- 컴포넌트 연결해주는 routes가 길어지면, 가독성이 좋지 않기 때문에 routes를 파일로 분리
- router 관련 config도 파일로 분리
