+++ client/src/app/router/config.ts
@@ -1,0 +1,9 @@
+export const routerConfig = {
+  future: {
+    v7_fetcherPersist: true,
+    v7_normalizeFormMethod: true,
+    v7_partialHydration: true,
+    v7_relativeSplatPath: true,
+    v7_skipActionErrorRevalidation: true,
+  },
+};
