+++ client/src/main.tsx
@@ -1,6 +1,6 @@
 import { createRoot } from 'react-dom/client';
 import './index.css';
-import { Router } from '@/pages/Router';
+import { Router } from '@/app/router';
 
 createRoot(document.getElementById('root')!).render(
   <>
